+++ src/ralph/api/__init__.py
@@ -1,4 +1,7 @@
 """Main module for Ralph's LRS API."""
+from functools import lru_cache
+from urllib.parse import urlparse
+
 import sentry_sdk
 from fastapi import Depends, FastAPI
 
@@ -8,6 +11,23 @@ from .. import __version__
 from .auth import AuthenticatedUser, authenticated_user
 from .routers import health, statements
 
+
+@lru_cache(maxsize=None)
+def get_health_check_routes():
+    """Return the health check routes."""
+    return [route.path for route in health.router.routes]
+
+
+def filter_transactions(event, hint):  # pylint: disable=unused-argument
+    """Filter transactions for Sentry."""
+    url = urlparse(event["request"]["url"])
+
+    if settings.SENTRY_IGNORE_HEALTH_CHECKS and url.path in get_health_check_routes():
+        return None
+
+    return event
+
+
 if settings.SENTRY_DSN is not None:
     sentry_sdk.init(
         dsn=settings.SENTRY_DSN,
@@ -15,6 +35,7 @@ if settings.SENTRY_DSN is not None:
         release=__version__,
         environment=settings.EXECUTION_ENVIRONMENT,
         max_breadcrumbs=50,
+        before_send_transaction=filter_transactions,
     )
 
 app = FastAPI()